login passes attempting to getMessageBYID to pass

+++ src/main/java/DAO/MessageDAO.java
@@ -40,20 +40,23 @@ public class MessageDAO {
     /*
      * deleting a message
      */
-    public boolean deleteMessageByMessageID(int message_id) {
+    public Message deleteMessageByMessageID(int message_id) {
         Connection connection = ConnectionUtil.getConnection();
         try {
             //Write SQL logic here
+            if (!messageExists(message_id)) {
+                return null;
+            }
             String sql = "DELETE FROM message WHERE message_id = ?";
             PreparedStatement preparedStatement = connection.prepareStatement(sql);
             preparedStatement.setInt(1, message_id);
-            int rowsAffected = preparedStatement.executeUpdate();
-            return rowsAffected > 0;
+            preparedStatement.executeUpdate();
+            return getMessageByID(message_id);
 
         }catch(SQLException e){
             System.out.println(e.getMessage());
         }
-        return false;
+        return null;
     }
     /*
      * inserting a message
@@ -94,8 +97,8 @@ public class MessageDAO {
      * get all messgages given username
      */
 
-     public Message getMessageByID(int message_id) {
-        Connection connection = ConnectionUtil.getConnection();
+    public Message getMessageByID(int message_id) {
+        Connection connection = ConnectionUtil.getConnection();      
         try {
             //Write SQL logic here
             String sql = "SELECT * FROM message WHERE message_id = ?";
@@ -103,20 +106,37 @@ public class MessageDAO {
             preparedStatement.setInt(1, message_id);
             ResultSet rs = preparedStatement.executeQuery();
             while(rs.next()){
-                Message message = new Message(
-                        rs.getInt("message_id"),
+                Message message = new Message(rs.getInt("message_id"),
                         rs.getInt("posted_by"),
                         rs.getString("message_text"),
                         rs.getLong("time_posted_epoch")
                         );
-                if (rs.getInt("message_id") == message_id) {
-                    return new Message(message.getMessage_id(), message.getPosted_by(), message.getMessage_text(), message.getTime_posted_epoch());
-                }
+                return message;
             }
         }catch(SQLException e){
             System.out.println(e.getMessage());
         }
         return null;
-     }
+    }
+    public boolean messageExists (int message_id) {
+        Connection connection = ConnectionUtil.getConnection();      
+        try {
+            String sql = "SELECT * FROM message WHERE message_id = ?";
+            PreparedStatement preparedStatement = connection.prepareStatement(sql);
+            preparedStatement.setInt(1, message_id);
+            ResultSet rs = preparedStatement.executeQuery();
+            if (rs.next()) {
+                Message message = new Message (
+                rs.getInt("message_id"),
+                rs.getInt("posted_by"),
+                rs.getString("message_text"),
+                rs.getLong("time_posted_epoch"));
+                return true;
+            }
+        } catch (SQLException e) {
+            System.out.println(e.getMessage());
+        }
+        return false;
+    }
 }
